extract jlcpcb cad model helper
